+++ tests/test_calendar.py
@@ -7,7 +7,7 @@ from datetime import date, datetime, timedelta
 
 # pylint: disable=import-error disable=wrong-import-position
 sys.path.append(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
-from pydit import create_calendar, setup_logging, fom_eom
+from pydit.functions import create_calendar, fom_eom, setup_logging
 
 
 # import numpy as np
@@ -22,14 +22,13 @@ def test_calendar():
     res = create_calendar(start="2022-01-01", end="2022-01-31")
     print(res)
     assert res.shape[0] == 31
-    assert res[res["date"] == "2022-01-24"]["weekday"].squeeze() == 0  # monday
+    assert res[res["date"] == "2022-01-24"]["weekday"].squeeze() == 1  # monday
     assert res[res["date"] == "2022-01-24"]["weekend"].squeeze() == False
     assert res[res["date"] == "2022-01-23"]["weekend"].squeeze() == True
     assert res[res["date"] == "2022-01-01"]["day_of_year"].squeeze() == 1
     assert (
         res[res["date"] == "2022-01-01"]["yyyyww"].squeeze() == 202152
     )  # ATTENTION THIS WILL WORK FOR SORTING WEEKS ON AXIS BUT NOT WEEK
-
 
 
 def test_fom_eom():
@@ -40,16 +39,10 @@ def test_fom_eom():
     assert fom_eom(date(2024, 8, 10), return_datetime=False)[0] == date(2024, 8, 1)
     assert fom_eom(date(2024, 8, 10), return_datetime=False)[1] == date(2024, 8, 31)
 
-    #example testing one more second after the end of the month
-    assert fom_eom(date(2024, 8, 10))[1] + timedelta(seconds=1) == datetime(2024, 9, 1, 0, 0, 0)
-
-
-
-
-
-
-
-
+    # example testing one more second after the end of the month
+    assert fom_eom(date(2024, 8, 10))[1] + timedelta(seconds=1) == datetime(
+        2024, 9, 1, 0, 0, 0
+    )
 
 
 if __name__ == "__main__":