+++ CHANGELOG.md
@@ -3,11 +3,19 @@
 ## V0.1.07
 
 ### Features
-  Added a new function to generate first and last day of the month given a date or datetime or string
-  Added this to the calendar creation function too.
+
+- Added a new function to generate first and last day of the month given a date or datetime or string. Added this to the calendar creation function too.
+- Added simulation class
 
 ### Fixes
+
   Removed inplace=True across the library to align to pandas direction, so now everything returns a new object and we don't mutate anything
+
+### Tweaks
+
+  Restructured the modules so we can have folder by types of function and importing them individually rather than full pydit module.
+
+  E.g. one can do from pydit.statistics import benford_to_dataframe
 
 ## V0.1.06